Optimized internal iterator reset on jump * LogicIterator offers its specializations the possibility of executing specific reset instructions before jumping ** this is currently used to reset internal optimizations of the LogicOrIterator and LogicXorIterator classes * The method was changed so that it now only resets if we jump backwards or if any of the 'done iterators' need to jump back ** old: void LogicIterator::resetInternal() ** new: void LogicIterator::resetBeforeJump(uint32_t)

+++ src/query/logic/logic_and_iterator.h
@@ -15,7 +15,7 @@ class LogicAndIterator : public LogicIterator {
 		virtual bool iteratorsCheckCommon();
 		std::vector<BasicIterator::Ptr>::iterator iteratorsGetMax();
 
-		uint32_t iterators_move_count_ = 0;
+		uint8_t iterators_move_count_ = 0;
 };
 
 }
